Fixed type casting in BatteryState.ino

+++ arduino/opencr_arduino/opencr/libraries/ROS2/examples/sensor_msgs/BatteryState/BatteryState.ino
@@ -50,8 +50,8 @@ private:
 
     sensor_msgs::BatteryState battery_state_topic;
     battery_state_topic.header.frame_id = (char*) "OpenCR BatteryState";
-    battery_state_topic.header.stamp.sec = nano_time_/1000000000;
-    battery_state_topic.header.stamp.nanosec = nano_time_%1000000000;
+    battery_state_topic.header.stamp.sec = (int32_t) (nano_time_/1000000000);
+    battery_state_topic.header.stamp.nanosec = (uint32_t) (nano_time_%1000000000);
     
     battery_state_topic.voltage = 1;
     battery_state_topic.current = 2;
